backend/requirements.txt
fastapi[all]
tortoise-orm
python-jose
fastapi-admin
